U/ryantrem/build engine options (#28)
Adding the ability to select a "build engine." By default, this is dotnet, but can be changed to Visual Studio.

In the case of OSX, selecting Visual Studio just changes the command to use msbuild instead of dotnet since Visual Studio for Mac adds msbuild to the path (untested).

In the case of Windows, selecting Visual Studio results in an invocation of the Visual Studio Installer to get the path to the latest version of Visual Studio and gets the msbuild path from it.

This change largely addresses issue #10. I simplified a little bit by just adding a single platform independent Visual Studio option (not separate options for Windows/Mac). If this turns out to not be sufficient, we can add more options later.