Enable interactive restore and ADO auth (#29)
This change adds -p:NuGetInteractive=true to the build command, and scans for ADO package feed authentication requests. The ADO auth request is logged from the build, and this is detected via a regex, but most of the string is localized so the regex is a bit different than might be expected. It looks for the non-localized prefix of [CredentialProvider], looks for the ADO auth url of https://microsoft.com/devicelogin, and looks for the 9 character auth code. A dialog is displayed that helps the user navigate to the auth page and enter the auth code.

Clicking OK will take the user to the authentication web page (https://microsoft.com/devicelogin).

After entering the code and clicking Next, the authentication will be completed and NuGet restore will complete and the build will continue.

This is arguably fragile because if the ADO NuGet auth plugin is updated and changes the format of the output, this could break. It would be more robust to just show the build output as multi-line selectable text and just let the user observe the auth request in the build output (same as would happen when building on the command line), but this actually requires a lot more code. Since the solution I have here is a small amount of code and yields a better workflow, I thought I'd start with this solution.

This addresses issue #4.