MSBuildForUnity NuGet package (#30)
Adding the MSBuildForUnity NuGet package, which is intended to augment MSBuildProjects that are built by the MSBuildForUnity ProjectBuilder (Unity integration). The package currently provides these three things:

- .meta file generation
- resolving the ProjectReference/PackageReference dependency graph and ensuring it all ends up in the output directory
- patching .pdbs to make pre-compiled dlls (e.g. from packages) debuggable when running the Unity editor

The major pieces of this change are:
- Add props/targets to augment the build as described above
- Add c# source for some custom MSBuild tasks used by the above
- Add support for building a NuGet package with the custom props/targets/tasks
- Add documentation for these new features
- Add yaml builds for generating the NuGet packages

For now, the NuGet package is publishing to a private test feed. Will sort this out later.