Update samples to be compliant with NuGet security analysis in Azure pipelines
Update samples to be compliant with automatic NuGet security analysis injected in ADO pipelines. See https://aka.ms/nugetsecurityanalysis for more details but this essentially boils down to:

1. Ensuring `<clear/>` statements are in each nuget.config file in the repo
2. nuget.config only specifies a singular nuget feed

Since these are samples, the change adds 1 above and comments out references to the MSBuildForUnity feed with the expectation that one would uncomment them from the sample when testing this out manually. This is currently consistent with the documentation that a special feed for the MSBuildForUnity is currently required.